+++ libk/string.c
@@ -61,12 +61,12 @@ void *strcpy(char *dest, const char *src)
   return memcpy(dest, src, strlen(src) + 1);
 }
 
-size_t strncmp(const char *s1, const char *s2, size_t n)
+int strncmp(const char *s1, const char *s2, size_t n)
 {
   return memcmp(s1, s2, n);
 }
 
-size_t strcmp(const char *s1, const char *s2)
+int strcmp(const char *s1, const char *s2)
 {
   size_t s1_len = strlen(s1);
   size_t s2_len = strlen(s2);
